Move game states to their own files
For #20
