+++ game/gamestates/__init__.py
@@ -1,0 +1,13 @@
+from .gamestate import GameState
+from .combatstate import CombatState
+from .characterselectionstate import CharacterSelectionState
+
+
+__all__ = ['states']
+
+
+#pylint: disable=invalid-name
+states = {
+    'Combat': CombatState,
+    'CharacterSelection': CharacterSelectionState,
+}
