+++ game/gamestates/gamestate.py
@@ -1,0 +1,17 @@
+from direct.showbase.DirectObject import DirectObject
+import panda3d.core as p3d
+
+
+class GameState(DirectObject):
+    def __init__(self):
+        super().__init__()
+        self.root_node = p3d.NodePath('State Root')
+        self.root_node.reparent_to(base.render)
+
+    def cleanup(self):
+        self.ignoreAll()
+        self.root_node.remove_node()
+        self.root_node = None
+
+    def update(self, _dt):
+        pass
